rename selector folder to smart-selector
