+++ libs/smart-ngrx/src/index.ts
@@ -4,13 +4,13 @@ export * from './common/for-next.function';
 export * from './facades/classic-ngrx.facade/action.factory';
 export * from './providers/provide-smart-feature-entities.function';
 export * from './providers/provide-smart-ngrx.function';
-export * from './selector/create-smart-selector.function';
+export * from './smart-selector/create-smart-selector.function';
 export * from './types/effect-service';
 export * from './types/row-proxy-delete.interface';
 export * from './types/smart-array.interface';
 
 // store.function is needed for testing
-export * from './selector/store.function';
+export * from './smart-selector/store.function';
 export * from './socket/handle-socket-notification.function';
 
 // testing utility functions (maybe move to their own lib?)
