+++ libs/smart-ngrx/src/row-proxy/force-refetch-of-virtual-indexes.function.spec.ts
@@ -1,8 +1,8 @@
 import { EntityState } from '@ngrx/entity';
 
 import { FacadeBase } from '../facades/facade.base';
-import { ArrayProxy } from '../selector/array-proxy.class';
-import { VirtualArray } from '../selector/virtual-array.class';
+import { ArrayProxy } from '../smart-selector/array-proxy.class';
+import { VirtualArray } from '../smart-selector/virtual-array.class';
 import { ChildDefinition } from '../types/child-definition.interface';
 import { SmartNgRXRowBase } from '../types/smart-ngrx-row-base.interface';
 import { VirtualArrayContents } from '../types/virtual-array-contents.interface';
