+++ libs/smart-ngrx/src/row-proxy/row-proxy.class.spec.ts
@@ -1,5 +1,5 @@
 import { FacadeBase } from '../facades/facade.base';
-import { ArrayProxy } from '../selector/array-proxy.class';
+import { ArrayProxy } from '../smart-selector/array-proxy.class';
 import { SmartNgRXRowBase } from '../types/smart-ngrx-row-base.interface';
 import { RowProxy } from './row-proxy.class';
