+++ libs/smart-ngrx/src/smart-selector/array-proxy-class.get.function.spec.ts
