+++ libs/smart-ngrx/src/smart-selector/convert-children-to-array-proxy.function.ts
