+++ libs/smart-ngrx/src/smart-selector/convert-children-to-virtual-array.function.ts
