+++ libs/smart-ngrx/src/smart-selector/create-inner-smart-selector.function.ts
