+++ libs/smart-ngrx/src/smart-selector/create-smart-selector.function.ts
