+++ libs/smart-ngrx/src/smart-selector/get-array-item.function.ts
