+++ libs/smart-ngrx/src/smart-selector/is-array-proxy.function.ts
