+++ libs/smart-ngrx/src/smart-selector/new-row-registry.class.ts
