+++ libs/smart-ngrx/src/smart-selector/parent-selector.type.ts
