+++ libs/smart-ngrx/src/smart-selector/real-or-mocked.function.spec.ts
