+++ libs/smart-ngrx/src/smart-selector/store.function.ts
