+++ libs/smart-ngrx/src/smart-selector/virtual-array-map.const.ts
