+++ libs/smart-ngrx/src/smart-selector/virtual-array.class.spec.ts
