+++ libs/smart-ngrx/src/smart-signals/create-smart-signal.function.ts--x
@@ -1,0 +1,24 @@
+import { Signal } from '@angular/core';
+import { Dictionary } from '@ngrx/entity';
+
+import { ChildDefinition } from '../types/child-definition.interface';
+import { SmartNgRXRowBase } from '../types/smart-ngrx-row-base.interface';
+
+/**
+ *
+ * @param parentSignal
+ * @param children
+ */
+export function createSmartSignal<
+  P extends SmartNgRXRowBase,
+  T extends SmartNgRXRowBase,
+>(parentSignal: Signal<Dictionary<P>>, children: ChildDefinition<P, T>[]): Signal<P> {
+  return children.reduce(createSmartSignalChildReducer<P, T>, parentSignal);
+}
+
+function createSmartSignalChildReducer<
+  P extends SmartNgRXRowBase,
+  T extends SmartNgRXRowBase,
+>(parentSignal: Signal<T>, childDefinition: ChildDefinition<P, T>) {
+  return createInnerSmartSignal(parentSignal, childDefinition.selector);
+}
