+++ libs/smart-ngrx/src/socket/update-entity.function.ts
@@ -5,7 +5,7 @@ import { forNext } from '../common/for-next.function';
 import { FacadeBase } from '../facades/facade.base';
 import { facadeRegistry } from '../registrations/facade-registry.class';
 import { featureRegistry } from '../registrations/feature-registry.class';
-import { store } from '../selector/store.function';
+import { store } from '../smart-selector/store.function';
 import { SmartNgRXRowBase } from '../types/smart-ngrx-row-base.interface';
 
 /**
