+++ libs/smart-ngrx/src/tests/functions/set-state.function.ts
@@ -2,7 +2,7 @@ import { MockStore } from '@ngrx/store/testing';
 import { take } from 'rxjs/operators';
 
 import { castTo } from '../../common/cast-to.function';
-import { store } from '../../selector/store.function';
+import { store } from '../../smart-selector/store.function';
 
 /**
  * Sets the state of the store during unit testing.  This grabs
