+++ libs/smart-ngrx/tsconfig.lib.json
@@ -16,5 +16,5 @@
     "jest.config.ts",
     "src/**/*.test.ts"
   ],
-  "include": ["src/**/*.ts"]
+  "include": ["src/**/*.ts", "src/smart-signals/create-smart-signal.function.ts--x"]
 }
